Update AdminDAO.java
Ik heb error handling toegevoegd aan het verwijderen van Admins zodat gebruikers nu duidelijk geinformeerd worden als het id dat ze willen verwijderen niet bestaat.

+++ src/main/java/nl/hsleiden/svdj8/daos/AdminDAO.java
@@ -39,6 +39,10 @@ public class AdminDAO {
     }
 
     public void deleteAdmin(long id) {
+        Optional<Admin> adminId = adminRepository.findById(id);
+        if (adminId.isEmpty()) {
+            throw new IllegalStateException("Delete failed, no admin with the id: " + id + " was found");
+        }
         adminRepository.deleteById(id);
     }
